move pinning state getting/setting inside the row model
